third_party/WebKit/LayoutTests/fast/css-grid-layout/grid-grow-tracks-to-their-max-expected.txt

Check that tracks are not shrunk if the available space is indefinite.
PASS window.getComputedStyle(gridContentSizedColumn, '').getPropertyValue('grid-template-columns') is "50px"
PASS window.getComputedStyle(gridContentSizedColumn, '').getPropertyValue('grid-template-rows') is "10px"
PASS window.getComputedStyle(gridContentSizedRow, '').getPropertyValue('grid-template-columns') is "40px"
PASS window.getComputedStyle(gridContentSizedRow, '').getPropertyValue('grid-template-rows') is "20px"


Check that tracks use the available free space if available.
PASS window.getComputedStyle(gridContentSizedColumnFreeSpace, '').getPropertyValue('grid-template-columns') is "80px"
PASS window.getComputedStyle(gridContentSizedColumnFreeSpace, '').getPropertyValue('grid-template-rows') is "10px"
PASS window.getComputedStyle(gridContentSizedRowFreeSpace, '').getPropertyValue('grid-template-columns') is "40px"
PASS window.getComputedStyle(gridContentSizedRowFreeSpace, '').getPropertyValue('grid-template-rows') is "80px"
PASS window.getComputedStyle(gridContentSizedColumnFreeSpaceParentDefinite, '').getPropertyValue('grid-template-columns') is "80px"
PASS window.getComputedStyle(gridContentSizedColumnFreeSpaceParentDefinite, '').getPropertyValue('grid-template-rows') is "10px"
PASS window.getComputedStyle(gridContentSizedRowFreeSpaceParentDefinite, '').getPropertyValue('grid-template-columns') is "40px"
PASS window.getComputedStyle(gridContentSizedRowFreeSpaceParentDefinite, '').getPropertyValue('grid-template-rows') is "100px"
PASS successfullyParsed is true

TEST COMPLETE

XXXXX
XXXX XXXX
XX XX
XXXX XXXX
XX XX
XXXX XXXX
